Rearranged code to validate form field before displaying confirmation message
